[Fix] Conflict when root and object property share the same name
