+++ spec/support/shared_examples/hash_examples.rb
@@ -1,0 +1,6 @@
+shared_examples 'hash is cloned deeply and it`s structure is not changed' do
+  specify do
+    expect(copy).to eq(object)
+    expect(copy).to eql(object)
+  end
+end
